refactor: menuButton extension 별도 파일로 분리, 종료된 classItem 관련 셀 변경사항 폐기, 주석정리
- menuButton extension 함수를 별도 파일로 분리하였습니다.
- 종료된 수업 관련 셀 변경사항을 폐기하였습니다.
- 간단한 주석정리를 진행했습니다.
